Fixes css animation not restarting after window loses focus
CSS animation events aren’t fired by browsers when the user navigates
to another tab or window (window.blur) and my code was relying on the
‘animationend’ event to restart the animation cycle. This worked well
when the user was focused on my app but breaks when focus is away.

Now resetting the animation duration and properties whenever
window.focus fires. A bit kludgy but it resolves the problem.

Also some cleanup and renaming for clarity and consistency.